% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{atGO2Potri}
\alias{atGO2Potri}
\title{Trace GO terms to gene names}
\usage{
atGO2Potri(go_id, mapping = MAPPING_tair10_GO_to_pt210)
}
\arguments{
\item{go_id}{GO term in format GO:0000000}

\item{mapping}{Specify a object to map GO to genes}
}
\value{
Returns a vector of genes that map to a GO term
}
\description{
This function maps GO terms back to gene names
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
