% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{atGO2PotriAnno}
\alias{atGO2PotriAnno}
\title{Trace GO terms to gene names and annotations}
\usage{
atGO2Potri(go_id, mapping = MAPPING_tair10_GO_to_pt210
annotation = Ptrichocarpa_210_annotation_primary, col_index=2)
}
\arguments{
\item{go_id}{GO term in format GO:0000000}

\item{mapping}{Specify a object to map GO to genes}

\item{annotation}{Specify a object to annotate genes}

\item{col_index}{Specify which column contains the gene names}
}
\value{
Returns a data frame of annotated genes
}
\description{
This function maps GO terms back to gene names and includes annotations
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
