% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{atGOanalysis}
\alias{atGOanalysis}
\title{GO enrichment analysis using Arabidopsis TAIR10}
\usage{
atGOanalysis(genes, universe = TAIR10_GO_Universe,
               organism = "Arabidopsis thaliana", pv = 0.01,
               ontology=c("BP","MF","CC"), testDirection = "over")
}
\arguments{
\item{genes}{Provide a vector of gene names to be tested}

\item{universe}{Arabidopsis TAIR10 GO universe}

\item{organism}{Specify the name of the organism}

\item{pv}{Specify a p-value threshhold for significance (default = 0.01)}

\item{ontology}{Specify the ontologies to test (Default = c("BP","MF","CC"))}

\item{testDirection}{A string which can be either "over" or "under". This determines whether the test performed detects over or under represented GO terms. See \link[Category]{GSEAGOHyperGParams} (Defalut = "over")}
}
\value{
Returns a list object that contains the ontology results for Biological Processes (BP), Molecular Functions (MF) and Cellular Components (CC)
}
\description{
This function does GO enrichment on TAIR10 GO universe
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
