% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GOanalysis}
\alias{GOanalysis}
\title{GO enrichment analysis}
\usage{
GOanalysis(genes, universe = pt210_GO_universe,
               organism = "Populus trichocarpa", pv = 0.01,
               ontology=c("BP","MF","CC"), testDirection = "over")
}
\arguments{
\item{genes}{Provide a vector of gene names to be tested}

\item{universe}{Provide a user define GO universe}

\item{organism}{Specify the name of the organism}

\item{pv}{Specify a p-value threshhold for significance (default = 0.01)}

\item{ontology}{Specify the ontologies to test (Default = c("BP","MF","CC"))}

\item{testDirection}{A string which can be either "over" or "under". This determines whether the test performed detects over or under represented GO terms. See \link[Category]{GSEAGOHyperGParams} (Defalut = "over")}
}
\value{
Returns a list object that contains the ontology results for Biological Processes (BP), Molecular Functions (MF) and Cellular Components (CC)
}
\description{
This function does GO enrichment on a user defined GO universe
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
