% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ParseGOBPnPlot}
\alias{ParseGOBPnPlot}
\title{Parse GO results using grep and plot}
\usage{
ParseGOBPnPlot(grepList,GOresults, minT=0, Rorder=NULL,
             my_palette=colorRampPalette(c("white", "red"))(n = 8),
             main = "GO Enrichment",
             zlim = c(0,15), hlines=NULL )
}
\arguments{
\item{grepList}{A list object where each value contains a group of search terms sperated by "|". The names and order values will be taken into account when plotting.}

\item{GOresults}{A list containing multiple GO$BP objects from GOanalysis}

\item{minT}{An integer specifying the minimum number of time a term must occur to be considered important. Default = 0}

\item{Rorder}{Row order for how the results shoul dbe plotted}

\item{main}{Title of the plot}

\item{zlim}{Range specifying the p-values to be plotted}

\item{hlines}{A `cumsum()` vector designating breaks for grouping modules}

\item{mypalette}{Color palette used for heatmap}
}
\value{
Returns summary figure and summary table of results used to generate figure
}
\description{
This function creates a summary plot of GO$BP results by parsing GO results using grep and a list of search terms
}
\examples{
#generate some data
     set1 <- sample(Ptrichocarpa_210_annotation_primary$ATG)[1:2000]
     set2 <- sample(Ptrichocarpa_210_annotation_primary$ATG)[1:2000]

#calculate GO enrichment for BP
     GOresults<-list
     GOresults$set1<-atGOanalysis(set1,ontology="BP)
     GOresults$set2<-atGOanalysis(set2,ontology="BP)

#create search list
     grepList<-list()
     grepList$hormone<-"hormone|gibberelli|brassino|auxin"
     grepList$peroxisome <- "peroxi"
     grepList$localization <-"protein localization"

results<-ParseGOBPnPlot(grepList, GOresults)

}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
