% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{removeTERM}
\alias{removeTERM}
\title{Remove GO term from GO_object}
\usage{
atGO2Potri(go_id, mapping = MAPPING_tair10_GO_to_pt210)
}
\arguments{
\item{GO_object}{A Go object from GOanalysis}

\item{go_id}{GO term in format GO:0000000}
}
\value{
Returns update GO_object
}
\description{
This function removes a GO term from GO_object
}
\author{
Matthew Zinkgraf, \email{mzinkgraf@gmail.com}
}
